% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{add_larvae}
\alias{add_larvae}
\title{Add missing larvae stages to the diet table.}
\usage{
add_larvae(diet_table)
}
\arguments{
\item{diet_table}{wide diet table.}
}
\value{
data.frame diet table with larvae stage.
}
\description{
Add missing larvae stages to the diet table.
}
