% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{add_observation_number}
\alias{add_observation_number}
\title{Add number of measured individuals to the catch data frame.}
\usage{
add_observation_number(valid_data)
}
\arguments{
\item{valid_data}{}
}
\value{
List of data frames.
}
\description{
Add number of measured individuals to the catch data frame.
}
