% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{amobio_selected_vars}
\alias{amobio_selected_vars}
\title{Environmental variables from the AMOBIO databsed selected for the analysis}
\usage{
amobio_selected_vars(average_time)
}
\arguments{
\item{average_time}{string indicating the duration of the average for
chemical variabels. Possible values are "1y" (one year), "3m" (three months),
"5y" (five years) and "all" (all available points).}
}
\value{
vector of string
}
\description{
Environmental variables from the AMOBIO databsed selected for the analysis
}
