% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{build_sem}
\alias{build_sem}
\title{Build a SEM.}
\usage{
build_sem(observable, mediator, response, random, data)
}
\arguments{
\item{observable}{list of observable variables}

\item{mediator}{list of mediator (intemediate) variables}

\item{response}{list of response variable that we want to predict}

\item{random}{random effects}

\item{data}{data frame}
}
\value{
psem output
}
\description{
Build a SEM.
}
