% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{clean_sea_data}
\alias{clean_sea_data}
\title{Clean sea data.}
\usage{
clean_sea_data(raw_data)
}
\arguments{
\item{raw_data}{output of \code{load_raw_sea_data}.}
}
\value{
List containing cleaned data frames for each sea campagne.
}
\description{
Clean sea data.
}
