% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_plot_dag}
\alias{create_plot_dag}
\title{Create the plot of the DAG using ggplot.}
\usage{
create_plot_dag()
}
\value{
ggplot
}
\description{
This represents the DAG we assume for our study.
That is, how environmental variable can shape food web structure.
}
