% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{download_amobio_data}
\alias{download_amobio_data}
\title{Download the AMOBIO data base}
\usage{
download_amobio_data(dir, verbose = FALSE)
}
\arguments{
\item{dir}{where the data is downloaded.}
}
\value{
nothing
}
\description{
From data.gouv at the link
https://entrepot.recherche.data.gouv.fr/dataset.xhtml?persistentId=doi:10.57745/EQYVLP
}
