% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{download_sea_data}
\alias{download_sea_data}
\title{Download sea survey data.}
\usage{
download_sea_data(data_folder)
}
\arguments{
\item{data_folder}{Where to store data files.}
}
\description{
Download sea survey data.
}
