% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_workshop_data}
\alias{download_workshop_data}
\title{Download the data from the spatiotemporal workshop from Zenodo.}
\usage{
download_workshop_data(dir)
}
\arguments{
\item{dir}{Directory where to store the zip file.}
}
\value{
Path of the downloaded zip file.
}
\description{
Download the data from the spatiotemporal workshop from Zenodo.
}
