% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{filter_out_rare}
\alias{filter_out_rare}
\title{Filter rare species from diet table.}
\usage{
filter_out_rare(diet_table, sea_data, occurence_min = 10)
}
\arguments{
\item{diet_table}{diet table wide}

\item{sea_data}{data list of sea surveys}

\item{occurence_min}{set to 10 by default, define when a species is rare}
}
\value{
data.frame diet table without rare species
}
\description{
Rare species are defined on occurences.
A species is rare if its occurence is lower than \code{occurence_min}.
}
