% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{filter_sampling_rephy}
\alias{filter_sampling_rephy}
\title{Filter rephy data keeping stations with non-small sampling effort}
\usage{
filter_sampling_rephy(rephy_data, sampling_min = 20)
}
\arguments{
\item{rephy_data}{data tibble}

\item{sampling_min}{minimum sampling to be kept}
}
\value{
data.frame filtered
}
\description{
Station with less samples than sampling_min are removed.
}
