% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{filter_surface_rephy}
\alias{filter_surface_rephy}
\title{Filter rephy data keeping samples above depth_max}
\usage{
filter_surface_rephy(rephy_data, depth_max = 1)
}
\arguments{
\item{rephy_data}{data.frame}

\item{depth_max}{double}
}
\value{
data.frame filtered
}
\description{
Filter rephy data keeping samples above depth_max
}
