% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{filter_valid}
\alias{filter_valid}
\title{Keep only species with valid name in data list.}
\usage{
filter_valid(data_list)
}
\arguments{
\item{data_list}{}
}
\value{
data_list where non-valid species have been removed.
}
\description{
Keep only species with valid name in data list.
}
