% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_year}
\alias{filter_year}
\title{Filter data such that year is in the specified time window}
\usage{
filter_year(data, year_min, year_max)
}
\arguments{
\item{data}{data tibble with \code{year} column.}

\item{year_min}{minimum year to be kept.}

\item{year_max}{maximal year to be kept.}
}
\value{
data tibble filtered.
}
\description{
Filter data such that year is in the specified time window
}
