% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{filter_year_rephy}
\alias{filter_year_rephy}
\title{Drop years with little from REPHY data}
\usage{
filter_year_rephy(data, year_min = 2000)
}
\arguments{
\item{data}{tibble.}

\item{year_min}{minimum year to be kept}
}
\value{
data tibble filtered.
}
\description{
Drop years with little from REPHY data
}
