% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_date_to_year_month}
\alias{from_date_to_year_month}
\title{Convert date to year and month}
\usage{
from_date_to_year_month(data, keep = FALSE)
}
\arguments{
\item{data}{tibble with \code{date} column.}

\item{keep}{logical, whether or not to keep the original \code{date} column.}
}
\value{
tibble
}
\description{
Convert date to year and month
}
