% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_connectance}
\alias{get_connectance}
\title{Compute the connectance of the web.}
\usage{
get_connectance(web)
}
\arguments{
\item{web}{}
}
\value{
numeric
}
\description{
Compute the connectance of the web.
}
