% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_diet_category}
\alias{get_diet_category}
\title{Get fish diet with standardized categories.}
\usage{
get_diet_category(species)
}
\arguments{
\item{species}{A character vector of species names.}
}
\value{
A data frame with species name ("species"), predator stage ("stage")
and prey category ("prey_category").
}
\description{
Get fish diet with standardized categories.
}
