% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_diet_category_wide}
\alias{get_diet_category_wide}
\title{Get fish diet with standardized categories in a wide format.}
\usage{
get_diet_category_wide(species)
}
\arguments{
\item{species}{List of species names.}
}
\value{
Same as get_diet_categery, but in wide format.
}
\description{
Get fish diet with standardized categories in a wide format.
}
