% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_frac_piscivorous}
\alias{get_frac_piscivorous}
\title{Fraction of piscivorous fishes.}
\usage{
get_frac_piscivorous(web, resource)
}
\arguments{
\item{web}{adjacency matix}

\item{resource}{list of resources so we exclude them when computing the
fraction.}
}
\value{
numeric
}
\description{
Fraction of piscivorous fishes.
}
