% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_life_stage.R
\name{get_life_stage_length}
\alias{get_life_stage_length}
\title{Get life stage length for species.}
\usage{
get_life_stage_length(life_stage_file, maturity_length, epsilon = 1e-06)
}
\arguments{
\item{life_stage_file}{path to the file}

\item{maturity_length}{data frame with maturity length for each species
extracted from fishbase}
}
\value{
dataframe
}
\description{
To get life stage length we use data provided by Anik to get lg0max, lg1max
and length at maturity.
When the data is missing, we use fishbase which can provide length at
maturity.
}
