% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_life_stage.R
\name{get_maturity_length}
\alias{get_maturity_length}
\title{Get maturity length using fishbase}
\usage{
get_maturity_length(species)
}
\arguments{
\item{species}{list of species}
}
\value{
data frame of species and their maturity length.
}
\description{
Get maturity length using fishbase
}
