% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_piscivorous}
\alias{get_piscivorous}
\title{Get the list of piscivorous species from the diet table.}
\usage{
get_piscivorous(diet_table)
}
\arguments{
\item{diet_table}{}
}
\value{
vector of piscivorous species
}
\description{
Get the list of piscivorous species from the diet table.
}
