% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_predation_window}
\alias{get_predation_window}
\title{Create predation window table.}
\usage{
get_predation_window(diet_table)
}
\arguments{
\item{diet_table}{}
}
\value{
data frame
}
\description{
Beta_min and beta_max corresponds respectively to the minimal and maximal
size on which species can feed on, relative to its own body size.
For now, and for simplicity, we take constant values for piscivorous species.
}
