% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_richness}
\alias{get_richness}
\title{Compute the species richness of food web.}
\usage{
get_richness(web)
}
\arguments{
\item{web}{}
}
\value{
numeric
}
\description{
Compute the species richness of food web.
}
