% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_river_size}
\alias{get_river_size}
\title{Get individual fish sizes from the ASPE river survey data.}
\usage{
get_river_size(individual_fish_file)
}
\arguments{
\item{individual_fish_file}{path to \code{output_individual_fish.rda}.}
}
\value{
tibble with columns operation_id, batch_id, length (cm),
latin_name, species_valid, is_valid. Can be passed directly to
\code{get_metaweb()}.
}
\description{
Get individual fish sizes from the ASPE river survey data.
}
