% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{get_sem_net_effects}
\alias{get_sem_net_effects}
\title{Compute net effects of observables on responses.}
\usage{
get_sem_net_effects(model)
}
\arguments{
\item{model}{piecewiseSEM model}
}
\value{
data frame
}
\description{
Compute net effects of observables on responses.
}
