% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{get_size_extrema}
\alias{get_size_extrema}
\title{Get minimal and maximal size for each fish species}
\usage{
get_size_extrema(data_size)
}
\arguments{
\item{data_size}{}
}
\value{
data frame of size extrema
}
\description{
Get minimal and maximal size for each fish species
}
