% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_species_list}
\alias{get_species_list}
\title{Get the species list of the data set.}
\usage{
get_species_list(data_list)
}
\arguments{
\item{data_list}{}
}
\value{
species list
}
\description{
Assume that species names are given by 'species_valid' column.
}
