% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_species_richness}
\alias{get_species_richness}
\title{Compute the species richness of the food web}
\usage{
get_species_richness(web)
}
\arguments{
\item{web}{matrix}
}
\value{
numeric
}
\description{
Compute the species richness of the food web
}
