% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_trophic_length}
\alias{get_trophic_length}
\title{Get trophic length from a food web adjacency matrix}
\usage{
get_trophic_length(web, tl_method = "ChainAveragedTL")
}
\arguments{
\item{web}{adjacency matrix of type matrix.}

\item{tl_method}{string, cheddar method to compute trophic levels.}
}
\value{
numeric, trophic length, that is the maximal trophic level of the
community.
}
\description{
Get trophic length from a food web adjacency matrix
}
