% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_trophic_richness}
\alias{get_trophic_richness}
\title{Compute the trophic richness of food web.}
\usage{
get_trophic_richness(web)
}
\arguments{
\item{web}{}
}
\value{
numeric
}
\description{
Trophic richness count class sizes of the same species as different trophic
species, so it is not always equal to the species richness.
}
