% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{impute_size}
\alias{impute_size}
\title{Impute missing fish length.}
\usage{
impute_size(data)
}
\arguments{
\item{data}{}
}
\value{
list of data.frame.
}
\description{
Infer missing missing fish size using a truncated normal distribution.
Mean and variance are given by the observed length distribution
of the species in the sample.
If there three or less observations the variance is set to zero.
Extrema of the truncated distribution are given by the species length
extrema observed the entire dataset.
}
