% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{infer_missing_size}
\alias{infer_missing_size}
\title{Entire pipeline to infer missing sizes from the tidy data list.}
\usage{
infer_missing_size(tidy_data)
}
\arguments{
\item{tidy_data}{}
}
\value{
list of data.frame.
}
\description{
Entire pipeline to infer missing sizes from the tidy data list.
}
