% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{join_amobio_aspe}
\alias{join_amobio_aspe}
\title{Join AMOBIO and ASPE datasets}
\usage{
join_amobio_aspe(amobio, aspe)
}
\arguments{
\item{amobio}{dataframe}

\item{aspe}{dataframe}
}
\value{
dataframe
}
\description{
Join AMOBIO and ASPE datasets
}
