% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{land_polygon}
\alias{land_polygon}
\title{Get land polygons (islands included) from rnaturalearth package.}
\usage{
land_polygon()
}
\value{
sf polygons, WGS84.
}
\description{
Get land polygons (islands included) from rnaturalearth package.
}
