% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{load_raw_sea_data}
\alias{load_raw_sea_data}
\title{Load raw sea survey data.}
\usage{
load_raw_sea_data(data_folder)
}
\arguments{
\item{data_folder}{Path to data folder.}
}
\value{
A nested list with one element per survey ("pomet", "nurse", "solper").
Each survey contains two data frames: \code{catch} and \code{size}.
}
\description{
Load raw sea survey data.
}
