% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{merge_operation}
\alias{merge_operation}
\title{Combine sea and river operation metadata into one table.}
\usage{
merge_operation(sea_operation, river_operation)
}
\arguments{
\item{sea_operation}{sea operation metadata, as in \code{sea_data_tidy$trait}.}

\item{river_operation}{river operation metadata, as returned by
\code{get_river_operation()}.}
}
\value{
combined tibble with columns operation_id, year, month, longitude,
latitude, survey.
}
\description{
Combine sea and river operation metadata into one table.
}
