% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{plot_amobio_metric}
\alias{plot_amobio_metric}
\title{Utility function to visualise AMOBIO data.}
\usage{
plot_amobio_metric(
  amobio_data,
  metric,
  title = NULL,
  subtitle = NULL,
  legend = NULL,
  log_transform = FALSE
)
}
\arguments{
\item{amobio_data}{dataframe}

\item{metric}{string specifying the metric name to plot}

\item{title}{string giving the title of the plot}

\item{subtitle}{string giving the subtitle of the plot}

\item{legend}{string giving the legend (color caption) of the plot}

\item{log_transform}{whether or not to log transform the metric}
}
\value{
plot
}
\description{
Utility function to visualise AMOBIO data.
}
