% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_data_on_map}
\alias{plot_data_on_map}
\title{Plot data points on the France map}
\usage{
plot_data_on_map(data)
}
\arguments{
\item{data}{data frame}
}
\value{
plot by ggplot
}
\description{
Plot data points on the France map
}
