% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{plot_rephy_map}
\alias{plot_rephy_map}
\title{Map REPHY sampling sites}
\usage{
plot_rephy_map(data, facet = TRUE)
}
\arguments{
\item{data}{tibble as returned by \code{\link[=extract_nutrients_rephy]{extract_nutrients_rephy()}}.}

\item{facet}{logical. Facet the map by \code{parameter_code}.}
}
\value{
ggplot
}
\description{
Plots distinct sampling sites (deduplicated on coordinates) over the
French coastline, to get a sense of the spatial resolution of the REPHY
network. When \code{facet} is \code{TRUE} (default), sites are faceted by nutrient
parameter to compare spatial coverage across parameters - some sites may
only report a subset of the nutrients of interest.
}
