% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{plot_rephy_temporal_coverage}
\alias{plot_rephy_temporal_coverage}
\title{Plot yearly observation counts of REPHY nutrient data}
\usage{
plot_rephy_temporal_coverage(data)
}
\arguments{
\item{data}{tibble as returned by \code{\link[=extract_nutrients_rephy]{extract_nutrients_rephy()}}.}
}
\value{
ggplot
}
\description{
Heatmap of the number of observations for each nutrient parameter, across
years. Useful to spot parameters or periods with poor sampling coverage.
}
