% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{plot_rephy_temporal_trend}
\alias{plot_rephy_temporal_trend}
\title{Plot yearly nutrient concentration trends by DCE district}
\usage{
plot_rephy_temporal_trend(data)
}
\arguments{
\item{data}{tibble as returned by \code{\link[=summarise_rephy_yearly]{summarise_rephy_yearly()}}.}
}
\value{
ggplot
}
\description{
Plot yearly nutrient concentration trends by DCE district
}
