% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_sampling}
\alias{plot_sampling}
\title{Plot sampling station sea and river survey.}
\usage{
plot_sampling(sea_data, station_file)
}
\arguments{
\item{sea_data}{tidy data frame of sea data.}

\item{station_file}{file containing information about river station.}
}
\value{
plot.
}
\description{
Plot sampling station sea and river survey.
}
