% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{plot_sea_mesh}
\alias{plot_sea_mesh}
\title{Plot a sea mesh, zoomed to a bounding box.}
\usage{
plot_sea_mesh(sea_mesh, mesh_edges, xlim, ylim, title = NULL)
}
\arguments{
\item{sea_mesh}{list(mesh, in_water, land), e.g. \code{build_sea_mesh()}'s
output.}

\item{mesh_edges}{sf linestrings, e.g. \code{sea_mesh_edges()}'s output.}

\item{xlim, ylim}{numeric length-2, plot extent (degrees).}

\item{title}{plot title.}
}
\value{
ggplot.
}
\description{
Plot a sea mesh, zoomed to a bounding box.
}
