% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{preprocess_sea_data}
\alias{preprocess_sea_data}
\title{Entire pipeline to prepare the sea data survey for analysis.}
\usage{
preprocess_sea_data(dir, solper_reftax)
}
\arguments{
\item{dir}{Directory of raw data.}
}
\value{
list of data.frame $catch and $size.
}
\description{
Pipe utility functions to process raw sea data.
}
