% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{remove_rare_river_size}
\alias{remove_rare_river_size}
\title{Remove rare species from the river size data frame.}
\usage{
remove_rare_river_size(river_size, occurence_min)
}
\arguments{
\item{river_size}{tibble as returned by \code{get_river_size()}.}

\item{occurence_min}{numeric threshold defining a rare species.}
}
\value{
river_size, trimmed of rare and unvalidated species.
}
\description{
Remove rare species from the river size data frame.
}
