% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{rename_amobio}
\alias{rename_amobio}
\title{Dictionnary of renamed columns in the AMOBIO database.}
\usage{
rename_amobio()
}
\value{
named vector
}
\description{
To keep track of the renaming and ensure consistency.
}
