% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rephy.R
\name{summarise_rephy_temporal_coverage}
\alias{summarise_rephy_temporal_coverage}
\title{Summarise yearly observation counts of REPHY nutrient data}
\usage{
summarise_rephy_temporal_coverage(data)
}
\arguments{
\item{data}{tibble as returned by \code{\link[=extract_nutrients_rephy]{extract_nutrients_rephy()}}.}
}
\value{
tibble with one row per (year, parameter).
}
\description{
Summarise yearly observation counts of REPHY nutrient data
}
