% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_sf}
\alias{to_sf}
\title{Convert a dataframe to sf object}
\usage{
to_sf(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
sf object
}
\description{
Assumes coordinates are given by the longitude and latitude columns.
}
