% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{transform_solper_name}
\alias{transform_solper_name}
\title{Transform solper code to scientific species name}
\usage{
transform_solper_name(df, fname, solper = FALSE)
}
\arguments{
\item{df}{data frame with species code names}

\item{fname}{path to solper reftax file}

\item{solper}{whether or not to include solper survey data}
}
\value{
data frame with species scientifc names
}
\description{
Transform solper code to scientific species name
}
