% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{widen_diet_category}
\alias{widen_diet_category}
\title{Widen the dataframe of species diets.}
\usage{
widen_diet_category(diet_category)
}
\arguments{
\item{diet_category}{Output of get_diet_category.}
}
\value{
Same as get_diet_categery, but in wide format.
}
\description{
Widen the dataframe of species diets.
}
