% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{clean_name}
\alias{clean_name}
\title{Clean species name of a dataframe using fishbase.}
\usage{
clean_name(df)
}
\arguments{
\item{df}{}
}
\value{
df with cleaned name.
}
\description{
Clean species name of a dataframe using fishbase.
}
