% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{combine_amobio_data}
\alias{combine_amobio_data}
\title{Combine the AMOBIO tibble in a single one.}
\usage{
combine_amobio_data(metrics, nodes)
}
\arguments{
\item{metrics}{tibble}

\item{nodes}{tibble}
}
\value{
sf object CRS Lambert-93
}
\description{
Also select variables of interest.
}
