% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_metaweb}
\alias{get_metaweb}
\title{Build the metaweb representing all possible interactions.}
\usage{
get_metaweb(
  size_table,
  diet_fish,
  diet_resource,
  predation_window,
  num_classes = 5,
  selected_resources = c("zooplankton", "phytoplankton", "biofilm", "zoobenthos",
    "macrophyte", "detritus")
)
}
\arguments{
\item{size_table}{data frame of each individual fish size and the operation
id.}

\item{diet_fish}{table of fish diet.}

\item{diet_resource}{table of resource diet.}

\item{predation_window}{table of fish predation window, defining the minimal
and maximal size at which they can eat (for piscivorous).}

\item{num_classes}{number of size classes used to build the metaweb, default
set to 5.}

\item{selected_resources}{which resources are considered at the base of the
web, by default we consider: zooplankton, phytoplankton, biofilm, zoobenthos,
macrophyte and detritus.}
}
\value{
metaweb as a matrix.
}
\description{
Build the metaweb representing all possible interactions.
}
