% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_food_web.R
\name{get_river_operation}
\alias{get_river_operation}
\title{Get metadata for each ASPE fishing operation.}
\usage{
get_river_operation(individual_fish_file)
}
\arguments{
\item{individual_fish_file}{path to \code{output_individual_fish.rda}.}
}
\value{
tibble with columns operation_id, year, longitude, latitude.
}
\description{
Get metadata for each ASPE fishing operation.
}
