% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_life_stage.R
\name{merge_diet_size}
\alias{merge_diet_size}
\title{Fill diet table with corresponding life stage size}
\usage{
merge_diet_size(diet, maturity_length)
}
\arguments{
\item{diet}{diet table}

\item{maturity_length}{table of species maturity length}
}
\value{
data.frame
}
\description{
Fill diet table with corresponding life stage size
}
