% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_venn_diagram}
\alias{plot_venn_diagram}
\title{Plot venn diagram of common species between surveys}
\usage{
plot_venn_diagram(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plot
}
\description{
Plot venn diagram of common species between surveys
}
