% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{remove_rare_sea_catch}
\alias{remove_rare_sea_catch}
\title{Remove rare species from sea data catch data frame}
\usage{
remove_rare_sea_catch(data, occurence_min)
}
\arguments{
\item{data}{list of data tibble with a catch element, which has itself a
\code{species_valid} and \verb{trait columns}.}

\item{ocurence_min}{numeric threshold defining a rare species.}
}
\value{
list of data tible with catch and size elements that has been trimmed
of rare species.
}
\description{
A species is considered to be rare if an occurence equal or inferior to
\code{occurence_min}.
}
