% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amobio.R
\name{rename_amobio}
\alias{rename_amobio}
\title{Rename selected environmental variables of the AMOBIO database.}
\usage{
rename_amobio(average_time = "1y")
}
\value{
vector of strings.
}
\description{
Rename selected environmental variables of the AMOBIO database.
}
