% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_workshop}
\alias{unzip_workshop}
\title{Unzip data from the spatiotemporal workshop.}
\usage{
unzip_workshop(zip_file, dir)
}
\arguments{
\item{zip_file}{path of the zip file.}

\item{dir}{directory where to store the unzipped files.}
}
\value{
Path of the unzipped directory.
}
\description{
Data is stored within zenodo folder.
}
