% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_sea_data.R
\name{validate_species_names}
\alias{validate_species_names}
\title{Validate species scientific names.}
\usage{
validate_species_names(data_list)
}
\arguments{
\item{data_list}{}
}
\value{
List of cleaned data frame.
}
\description{
For now, we remove all species who have not a name validated by fishbase.
}
